ci: Run on pull requests (#27)
Right now the CI is configured to only run on pushes (to this repository), but not on PRs (from another repository).

The downside of this is that PRs for a branch in this repository will run twice. If you only push directly to e.g. master, it might be worth adjusting that - also see https://github.community/t/how-to-trigger-an-action-on-push-or-pull-request-but-not-both/16662